1- Remove space in the word page first line. 2-Add attendant name in title in new page if word reached the end of the page by 4 lines 3- Kashida for "الرئيس" , "الامين العام" 4- Add topics new feature 5- Add votes table 6- Increase attachment resolution 7- Partition word by each attachment in separate files to avoid word crash 8- Save html to word by using word procession 9- Reformat the file start 10- Add Attendant degree - order attanedant by name then insertion date 11-Dsiable edit "النصاب"  if the session file are under processing 12- Adjust the page margins and adjust the line numbers on word according to the new margins

+++ TayaIT.Enterprise.EMadbatah.Word/WordDocument.cs
@@ -52,7 +52,7 @@ namespace TayaIT.Enterprise.EMadbatah.Word
             return wordApp.GetNumPages();
         }
 
-        public long GetCurrentPageLineNumber(bool end)
+        public int GetCurrentPageLineNumber(bool end)
         {
             return wordApp.GetCurrentPageLineNumber(end);
         }
